move streaming models up to root of openapi
