+++ core/src/main/java/ru/tinkoff/invest/openapi/StreamingContext.java
@@ -3,8 +3,8 @@ package ru.tinkoff.invest.openapi;
 import org.jetbrains.annotations.NotNull;
 import org.reactivestreams.Publisher;
 
-import ru.tinkoff.invest.openapi.models.streaming.StreamingEvent;
-import ru.tinkoff.invest.openapi.models.streaming.StreamingRequest;
+import ru.tinkoff.invest.openapi.streaming.StreamingEvent;
+import ru.tinkoff.invest.openapi.streaming.StreamingRequest;
 
 public interface StreamingContext {
     void sendRequest(@NotNull StreamingRequest request);
