+++ core/src/main/java/ru/tinkoff/invest/openapi/streaming/StreamingRequest.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.invest.openapi.models.streaming;
+package ru.tinkoff.invest.openapi.streaming;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
