+++ example/src/main/java/ru/tinkoff/invest/openapi/example/StreamingApiSubscriber.java
@@ -6,7 +6,7 @@ import java.util.logging.Logger;
 import org.jetbrains.annotations.NotNull;
 import org.reactivestreams.example.unicast.AsyncSubscriber;
 
-import ru.tinkoff.invest.openapi.models.streaming.StreamingEvent;
+import ru.tinkoff.invest.openapi.streaming.StreamingEvent;
 
 class StreamingApiSubscriber extends AsyncSubscriber<StreamingEvent> {
